Fix overflow in TokenProgram Transfer by changing Amount to Uint64
The Amount field in the TokenProgram Transfer instruction was previously
using Uint32, which caused overflow when handling larger values. To fix
this, the Amount field has been updated to Uint64, providing sufficient
precision and preventing overflow issues during transfers.

+++ solana/parser_test.go
@@ -223,7 +223,7 @@ func TestTokenProgramTransfer(t *testing.T) {
 		assert.Equal(t, transferAction.InstructionName, "Transfer")
 		assert.Equal(t, transferAction.From, "6MkxxzHTzR9KJfd7PSr1c47ufbY3mfPgyqfgJQso7vtL")
 		assert.Equal(t, transferAction.To, "3ZiTyuCBHqFocKpFvU8wPwd6YM284aswxMZfad4w2ode")
-		assert.Equal(t, transferAction.Amount, uint32(3000000000))
+		assert.Equal(t, transferAction.Amount, uint64(3000000000))
 	} else {
 		t.Errorf("Error type assertion")
 	}
